PUSHING TO PC!
Pushing some minor changes back to PC after trying to debug on the mac. The error that I kept facing on my computer was due to the vertical colorbar that I am trying to implement to show the gradient shift in magnetic field strength. Hopefully can get that working soon but right now trying to fix the heatmaps as a whole since they're in a pretty meh state.