Text, Clusters, and Advanced Mapping
Updated the code for the color bar so that text shows up for right color bar, and also contains the unit (nT). Scatterplots with cluster identification was completed for all migratory birds, and their respective files are included in this update. Mapping improvements made as well, increasing efficiency and reducing rendering times.