MAJOR UPDATES. THE TREK IS OVER.
changed massproduction.py and am running the code on my computer overnight to hopefully get something running. it would be a miracle if I could analyze all of this data and add it into my paper! we'll see if my custom chunking method works out and if the file can be processed.